// vue/src/types/logger.ts
export type HrTime = [number, number];
export type TimeInput = HrTime | number | Date;
export enum SeverityNumber {
	UNSPECIFIED = 0,
	TRACE = 1,
	TRACE2 = 2,
	TRACE3 = 3,
	TRACE4 = 4,
	DEBUG = 5,
	DEBUG2 = 6,
	DEBUG3 = 7,
	DEBUG4 = 8,
	INFO = 9,
	INFO2 = 10,
	INFO3 = 11,
	INFO4 = 12,
	WARN = 13,
	WARN2 = 14,
	WARN3 = 15,
	WARN4 = 16,
	ERROR = 17,
	ERROR2 = 18,
	ERROR3 = 19,
	ERROR4 = 20,
	FATAL = 21,
	FATAL2 = 22,
	FATAL3 = 23,
	FATAL4 = 24,
}
export type AnyValueScalar = string | number | boolean;
export type AnyValueArray = Array<AnyValue>;
export interface AnyValueMap {
	[attributeKey: string]: AnyValue;
}
export type AnyValue = AnyValueScalar | Uint8Array | AnyValueArray | AnyValueMap | null | undefined;
export type LogBody = AnyValue;
export type LogAttributes = AnyValueMap;
export interface Context {
	getValue: (key: symbol) => unknown;
	setValue: (key: symbol, value: unknown) => Context;
	deleteValue: (key: symbol) => Context;
}
export interface LogRecord {
	eventName?: string;
	timestamp?: TimeInput;
	observedTimestamp?: TimeInput;
	severityNumber?: SeverityNumber;
	severityText?: string;
	body?: LogBody;
	attributes?: LogAttributes;
	context?: Context;
}